feat: PeekableStream 添加 trunc 接口 feat: 线程安全的 FixedBuffer feat: Socket 支持 PeekableStream 接口 docs: 补充部分文档

+++ sese/src/net/Socket.cpp
@@ -10,11 +10,11 @@ int sese::net::Socket::listen(socket_t socket, int backlog) noexcept {
 
 #ifdef SESE_PLATFORM_WINDOWS
 
-size_t sese::net::Socket::read(socket_t socket, void *buffer, size_t len, int flags) noexcept {
+int64_t sese::net::Socket::read(socket_t socket, void *buffer, size_t len, int flags) noexcept {
     return ::recv(socket, (char *) buffer, (int) len, flags);
 }
 
-size_t sese::net::Socket::write(socket_t socket, const void *buffer, size_t len, int flags) noexcept {
+int64_t sese::net::Socket::write(socket_t socket, const void *buffer, size_t len, int flags) noexcept {
     return ::send(socket, (const char *) buffer, (int)len, flags);
 }
